+++ django_fido/js/fido2.js
@@ -14,6 +14,7 @@ function createTranslations() {
         'ConstraintError': gettext('The security token (FIDO 2) does not meet required criteria.'),
         'NotAllowedError': gettext('The request has been refused either by you, your browser, or your'
                                    + ' operating system. Or the request timed out.'),
+        'NoAuthenticatorsError': gettext('Can\'t create FIDO 2 authentication request, no authenticators found.'),
     }
     TRANSLATIONS.GENERIC_ERROR_MESSAGE = gettext('An unknown error has occurred.')
     TRANSLATIONS.FIDO2_NOT_AVAILABLE = gettext('FIDO 2 is not available. Your browser may not support it'
@@ -125,7 +126,12 @@ async function sendFido2Request(url, is_registration, credenetials_name, form_da
             fido2ErrorResponseCallback(error)
         }
     } else {
-        fido2ErrorResponseCallback(null)
+        try {
+            const error_response = await response.json()
+            fido2ErrorResponseCallback({name: error_response.error_code})
+        } catch (error) {
+            fido2ErrorResponseCallback(null)
+        }
     }
 }