+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog #
 
+## Unreleased ##
+ * Display no authenticator error
+ * `error` key in repsponse of `BaseFido2RequestView` is now deprecated and will be removed in the future
+
 ## 0.22 ##
  * Use username as a backup `displayName`